+++ Jama/src/main/java/daoLayer/DepartmentDaoBean.java
@@ -8,31 +8,27 @@ import businessLayer.Department;
 
 @Stateless
 public class DepartmentDaoBean {
-	
-	@PersistenceContext(unitName="primary")
+
+	@PersistenceContext(unitName = "primary")
 	private EntityManager em;
 
 	public DepartmentDaoBean() {
 	}
-	
-	public Department createDepartment(String code, String name) {
-		
-		Department department = new Department();
-		department.setCode(code);
-		department.setName(name);
+
+	public Department createDepartment(Department department) {
+
 		em.persist(department);
-	
+
 		return department;
-		}
-	
-	public void removeDepartment(int id){
-		
+	}
+
+	public void removeDepartment(int id) {
+
 		Department department = em.find(Department.class, id);
-		em.remove(department);
-		
-		
-	}
-		
+		if (department != null) {
+			em.remove(department);
+		}
+
 	}
 
-
+}
